+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Add `textStyle` prop to `ToastMessage`
 * Add border to `Search` component
 
 ### Changed
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Increase top padding in `ChatMessage` content
+* Fix misalignment in replies text in `ChatMessage`
 
 ## [0.23.0] - 2022-04-29